Repasados todos los unit test para mejor cobertura

+++ src/ActivityHistory/Application/ActivityHistoryDeleteByNameService.php
@@ -28,7 +28,15 @@ class ActivityHistoryDeleteByNameService
 
     public function __invoke(array $data): JsonResponse
     {
+        if(!$this->projectRepository->checkIfProjectExists($data['projectName'])){
+            return new JsonResponse(['error' => 'Project with name ' . $data['projectName'] . ' was not found'], 404);
+        }
         $project = $this->projectRepository->findProjectByName($data['projectName']);
+
+        if(!$this->activityHistoryRepository->checkIfActivityHistoryFromProjectExists($data['name'], $project)){
+            return new JsonResponse(['error' => 'An activity with this name (' . $data['name'] . ') in project ' . $data['projectName'] . ' was not found'], 404);
+        }
+
         $activity = $this->activityHistoryRepository->findActivityHistoryFromProject($data['name'], $project);
         $this->activityHistoryRepository->removeActivityHistory($activity);
         return new JsonResponse(['message' => 'Activity deleted successfully'], 201);
